apps/backend/pyproject.toml
[project]
name = "image-promotion-backend"
version = "0.1.0"
description = "Backend API para demonstração de GitOps com promoção de imagens Docker"
authors = [{ name = "Marcell Martini" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.0,<5.0)",
]

[tool.poetry.group.dev.dependencies]
pytest="^8.0.0"
pytest-cov="^4.1.0"
pytest-asyncio="^0.23.0"
httpx="^0.27.0"
black="^24.0.0"
pylint="^3.0.0"
pyright="^1.1.0"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pylint.master]
init-hook = "import sys; sys.path.append('src')"

[tool.pylint.messages_control]
disable = ["C0114", "C0115", "C0116", "R0903", "R0801"]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "**/.poetry",
    "**/.venv",
]
extraPaths = ["src"]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["src/main.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pass",
    "@abstractmethod",
]
